mcp-servers/jenkins-mcp/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-jenkins==1.8.2
google-cloud-secret-manager==2.20.0
